Fix: deg

+++ WeatherApp.Infrastrucrure/OpenWeatherAPI/Entity/Wind.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 
 namespace WeatherApp.Infrastructure.OpenWeatherAPI.Entity
 {
-    public record Wind(decimal speed, int deg)
+    public record Wind(decimal speed, double deg)
     {
     }
 }
